requirements.txt
antlr4-python3-runtime==4.9.3
hydra-core==1.3.2
omegaconf==2.3.0
packaging==24.1
PyYAML==6.0.1
antlr4-python3-runtime==4.9.3
filelock==3.15.4
fsspec==2024.6.1
hydra-core==1.3.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
omegaconf==2.3.0
packaging==24.1
PyYAML==6.0.1
setuptools==71.1.0
sympy==1.13.1
torch==2.4.0
typing_extensions==4.12.2
